Add files via upload
Welcome to the Practical Data Science and Statistics Manual! This repository is designed to provide hands-on experience and practical knowledge in the fields of data science and statistics. Whether you're a beginner looking to get started or an experienced practitioner aiming to sharpen your skills, this manual has something for everyone.